requirements.txt
fastapi
uvicorn[standard]
openai
pinecone
requests
PyMuPDF
python-dotenv
groq
sentence-transformers
